+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * fix inserting emoji using Windows 10 emoji picker
 * fix position of textarea in multi line command (probably will affect IME or Emoji picker)
 * make jump to bottom on click only when terminal is not enabled [#596](https://github.com/jcubic/jquery.terminal/issues/596)
+* fix init cmd plugin without terminal
 
 ## 2.17.5
 ### Bugfix